Load xml.

+++ test/test_fx_rate_datasource.rb
@@ -2,27 +2,27 @@ require 'test/unit'
 require_relative '../lib/fx_rate_datasource'
 
 class TestFxRateDatasource < Test::Unit::TestCase
-  TEST_FEED_FILE = IO.readlines('eurofxref-hist-90d.xml')
-  # Called before every test method runs. Can be used
-  # to set up fixture information.
-  def setup
-    # Do nothing
-  end
-
-  # Called after every test method runs. Can be used to tear
-  # down fixture information.
+  TEST_FEED = 'eurofxref-hist-90d.xml'
 
   def teardown
-    File.delete(FxRateDatasource::FX_FEED_FILE)
-  end
-
-  def test_loads_fx_feed_to_local_file
-    FxRateDatasource.new.load_fx_rate('eurofxref-hist-90d.xml')
-    IO.readlines(FxRateDatasource::FX_FEED_FILE).each_with_index do |line, line_num|
-      puts "#{line_num}: #{line}"
-      assert_equal(TEST_FEED_FILE[line_num], line)
+    if File.exist?(FxRateDatasource::FX_FEED_FILE)
+      File.delete(FxRateDatasource::FX_FEED_FILE)
     end
   end
 
-  #def test
+  def test_loads_fx_feed_to_local_file
+    FxRateDatasource.new(TEST_FEED)
+    test_feed_file = IO.readlines(TEST_FEED)
+    #Compare expected and actual feed files, line by line
+    IO.readlines(FxRateDatasource::FX_FEED_FILE).each_with_index do |line, line_num|
+      assert_equal(test_feed_file[line_num], line)
+    end
+  end
+
+  #TODO Parameterise this
+  def test_returns_fx_rate
+    data = FxRateDatasource.new(TEST_FEED)
+    actual_fx_rate = data.get_fx_rate(Date.new(2017, 03, 02), 'GBP')
+    assert_equal(0.8556, actual_fx_rate)
+  end
 end
